added : Web config (fix CORS issues) and some HTTP methods support

+++ src/main/java/react/mongo/controller/GroupController.java
@@ -1,6 +1,5 @@
 package react.mongo.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import react.mongo.domain.Group;
 import react.mongo.domain.User;
@@ -8,13 +7,12 @@ import react.mongo.repository.GroupRepository;
 import react.mongo.repository.UserRepository;
 import react.mongo.utils.TransformUtils;
 
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 
 @RestController
 public class GroupController {
-
+    public static final String DEFAULT_GROUP_ID = "0bd01274-8a22-40bd-8131-a5b37a2be82f";
+    private static boolean isDefaultGroupExists = false;
     private final GroupRepository repository;
     private final UserRepository studentRepository;
     private final TransformUtils transformUtils;
@@ -23,16 +21,31 @@ public class GroupController {
         this.repository = repository;
         this.studentRepository = studentRepository;
         this.transformUtils = transformUtils;
+
+        if (!this.repository.existsById(DEFAULT_GROUP_ID)) {
+            Group defaultGroup = Group.builder()
+                    .key(DEFAULT_GROUP_ID)
+                    .name("Вне группы")
+                    .courseLevel("отсутствует")
+                    .direction("Оставшиеся юзеры")
+                    .users(new ArrayList<>())
+                    .build();
+
+            this.repository.save(defaultGroup);
+
+        }
+
+        isDefaultGroupExists = true;
     }
 
     @PostMapping("/groups")
-    public void saveGroup( @RequestBody LinkedHashMap<String, Object> entity) {
+    public void saveGroup(@RequestBody LinkedHashMap<String, Object> entity) {
         Group group = transformUtils.JSON2GroupTransform(entity, false);
         repository.save(group);
     }
 
     @PutMapping("/groups")
-    public void updateGroup( @RequestBody LinkedHashMap<String, Object> entity) {
+    public void updateGroup(@RequestBody LinkedHashMap<String, Object> entity) {
         Group group = transformUtils.JSON2GroupTransform(entity, true);
         repository.save(group);
     }
@@ -44,11 +57,24 @@ public class GroupController {
         Set<User> set = group.getUsers();
         set.add(student);
         group.setUsers(set);
+        List<Group> groupList = repository.findAll();
+        for(Group g : groupList){
+            if(g.getUsers().contains(student)){
+                Set<User> listUsers = g.getUsers();
+                listUsers.remove(student);
+                g.setUsers(listUsers);
+                repository.save(g);
+            }
+        }
         repository.save(group);
+
     }
 
     @DeleteMapping("/groups/{key}")
-    public void deleteGroup( @PathVariable String key ) {
+    public void deleteGroup(@PathVariable String key) {
+        if(key.equals(DEFAULT_GROUP_ID)){
+            return ;
+        }
         repository.deleteByKey(key);
     }
 
@@ -60,12 +86,13 @@ public class GroupController {
     }
 
     @DeleteMapping("/groups/{groupId}&{studentId}")
-    public void deleteStudentFromGroup(@PathVariable String groupId, @PathVariable String studentId){
+    public void deleteStudentFromGroup(@PathVariable String groupId, @PathVariable String studentId) {
         Group group = repository.findByKey(groupId);
         User student = studentRepository.findByKey(studentId);
         Set<User> set = group.getUsers();
         set.remove(student);
         group.setUsers(set);
+        addStudentToGroup(DEFAULT_GROUP_ID,student.getKey());
         repository.save(group);
     }
